@@ -1,7 +1,6 @@
 var UZIP = {};
 
-UZIP["parse"] = function (buf, onlyNames)	// ArrayBuffer
-{
+UZIP.parse = function (buf, onlyNames) {
     var rUs = UZIP.bin.readUshort, rUi = UZIP.bin.readUint, o = 0, out = {};
     var data = new Uint8Array(buf);
     var eocd = data.length - 4;
@@ -803,4 +802,4 @@ UZIP.F.U = function () {
     */
 })()
 
-if (typeof module !== 'undefined' && typeof module.exports !== 'undefined') module.exports = UZIP;
+export default UZIP;
